--- src/utils.ts
@@ -1,16 +1,0 @@
-import { streamers } from "./constants";
-
-export const decodeText = (text: string): string => {
-  return (
-    new DOMParser().parseFromString(decodeURI(text), "text/html")
-      .documentElement.textContent || ""
-  );
-};
-
-export const mergeFlag = (values: number[]): number => {
-  return values.reduce((acc, cur) => acc | cur, 0);
-};
-
-export const streamerFlagList = Object.values(streamers).map(
-  ({ flag }) => flag
-);
