+++ src/utils/decode.ts
@@ -1,0 +1,6 @@
+export const decodeText = (text: string): string => {
+  return (
+    new DOMParser().parseFromString(decodeURI(text), "text/html")
+      .documentElement.textContent || ""
+  );
+};
